add signed_url method to AWS::S3::File and tests
this returns the signed URL for public access to the file, the expiry
time can be passed into the method or will default to now + 1 hour in
the future. add perldoc and tests for these changes, update MANIFEST

although the URL is built on the client side this has been implemented
using the same structure as other requests with a GetPreSignedUrl class
that uses the AWS::S3:Roles::Request role to get at the _uri attribute

we don't actually go out to AWS for this method, but who knows what the
future holds?